test: move test for recalculate workflow tags to test_data
