+++ tests/samples/test_db.py
@@ -80,70 +80,6 @@ class TestCalculateWorkflowTags:
         assert tags == {"pathoscope": path_tag, "nuvs": nuvs_tag}
 
 
-class TestRecalculateWorkflowTags:
-    async def test(self, mocker, mongo):
-        await mongo.samples.insert_one(
-            {"_id": "test", "pathoscope": False, "nuvs": False}
-        )
-
-        analysis_documents = [
-            {
-                "_id": "test_1",
-                "workflow": "pathoscope_bowtie",
-                "ready": "ip",
-                "sample": {"id": "test"},
-            },
-            {
-                "_id": "test_2",
-                "workflow": "pathoscope_bowtie",
-                "ready": True,
-                "sample": {"id": "test"},
-            },
-            {
-                "_id": "test_3",
-                "workflow": "nuvs",
-                "ready": True,
-                "sample": {"id": "test"},
-            },
-        ]
-
-        await mongo.analyses.insert_many(
-            analysis_documents
-            + [
-                {
-                    "_id": "test_4",
-                    "sample": {"id": "foobar"},
-                    "workflow": "pathoscope_bowtie",
-                    "ready": True,
-                }
-            ],
-            session=None,
-        )
-
-        m = mocker.patch(
-            "virtool.samples.utils.calculate_workflow_tags",
-            return_value={"pathoscope": True, "nuvs": "ip"},
-        )
-
-        await recalculate_workflow_tags(mongo, "test")
-
-        for document in analysis_documents:
-            del document["sample"]
-
-        assert m.call_args[0][0] == analysis_documents
-
-        assert await mongo.samples.find_one() == {
-            "_id": "test",
-            "pathoscope": True,
-            "nuvs": "ip",
-            "workflows": {
-                "aodp": "incompatible",
-                "nuvs": "complete",
-                "pathoscope": "complete",
-            },
-        }
-
-
 class TestDeriveWorkflowStates:
     @pytest.mark.parametrize("library_type", ["amplicon", "normal", "srna", "other"])
     def test_define_initial_workflows(self, library_type, snapshot):
